requirements.txt
absl-py==0.11.0
atomicwrites==1.4.0
attrs==20.2.0
backcall==0.2.0
bitarray==1.6.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
einops @ file:///home/conda/feedstock_root/build_artifacts/einops_1602792478173/work
future==0.18.2
google-auth==1.23.0
google-auth-oauthlib==0.4.2
grpcio==1.33.2
hiddenlayer==0.3
idna==2.10
imageio==2.9.0
importlib-metadata==2.0.0
iniconfig==1.1.1
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
joblib==1.0.1
jsonpatch==1.31
jsonpointer==2.0
kiwisolver==1.2.0
libtiff==0.4.2
Markdown==3.3.3
matplotlib==3.3.2
networkx==2.5
numpy==1.19.2
oauthlib==3.1.0
opencv-python==4.4.0.44
packaging==20.4
parso==0.7.1
pickleshare==0.7.5
Pillow==8.0.0
pluggy==0.13.1
prompt-toolkit==3.0.8
protobuf==3.13.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.7.2
pylibtiff @ file:///C:/Users/98471/Downloads/pylibtiff-0.4.2-cp37-cp37m-win_amd64.whl
pyparsing==2.4.7
pytest==6.1.1
python-dateutil==2.8.1
PyWavelets==1.1.1
pyzmq==22.0.3
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.24.1
scipy==1.5.3
six==1.15.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.1
threadpoolctl==2.1.0
tifffile==2020.10.1
toml==0.10.1
torch @ file:///E:/zjh/torch-1.6.0%2Bcu101-cp37-cp37m-win_amd64.whl
torchfile==0.1.0
torchnet==0.0.4
torchsummary==1.5.1
torchvision @ file:///E:/zjh/torchvision-0.7.0%2Bcu101-cp37-cp37m-win_amd64.whl
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
urllib3==1.26.2
visdom==0.1.8.9
wcwidth==0.2.5
websocket-client==0.58.0
Werkzeug==1.0.1
wincertstore==0.2
zipp==3.4.0
